requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
torch>=1.8.0
torchvision>=0.9.0
onnxruntime>=1.8.0
tqdm
Pillow
